requirements.txt
pillow~=9.4.0
numpy~=1.24.3
pandas~=2.0.3
thop~=0.1.1-2209072238
tabulate~=0.9.0
scipy~=1.10.1
timm~=0.9.2
einops~=0.7.0
albumentations~=1.3.1